Use blueprint for application
