+++ app/admin/__init__.py
@@ -1,0 +1,7 @@
+# TODO: web backend for manage
+
+from flask import Blueprint
+
+admin = Blueprint('admin', __name__)
+
+from . import views, user
