+++ app/admin/views.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+# TODO: 管理后台视图
