+++ app/api_1_0/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+api = Blueprint('api', __name__)
+
+from . import admin, user
