+++ app/auth/views.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+# TODO: user login, logout, register, reset password ...
