+++ app/static/timer.js
