+++ app/templates/activated.html
