+++ app/templates/form_login.html
@@ -1,12 +1,12 @@
 {% extends 'form.html' %}
 {% from 'macros.html' import field, submit %}
 
-{% block title %}{{ _('Sign in - %(site)s', site=config['site_name']) }}{% endblock %}
+{% block title %}{{ _('Sign in - %(site)s'.format(site=config['site_name'])) }}{% endblock %}
 {% block body_class %}box_center{% endblock %}
 {% block form_title %}{{ _('Sign in') }}{% endblock %}
 
 {% block script %}
-var SUBMIT_URL = '{{ url_for('user_login') }}';
+var SUBMIT_URL = '{{ url_for('api.user_login') }}';
 var LOCK_FORM_AFTER_OK = false;
 var REDIRECT_URL = '/';
 var SUCCESS_CALLBACK = null;
