+++ app/validation.py
@@ -1,6 +1,4 @@
 #!/usr/bin/env python3
-
-
 import re
