+++ requirements/common.txt
@@ -1,0 +1,27 @@
+Flask==0.11.1
+Flask-Script==2.0.5
+Flask-Bootstrap==3.3.7.0
+Flask-Moment==0.5.1
+Flask-WTF==0.13
+Flask-SQLAlchemy==2.1
+Flask-WhooshAlchemy==0.56
+Flask-WhooshAlchemyPlus==0.7.5
+Flask-Migrate==2.0.0
+Flask-Mail==0.9.1
+Flask-Login==0.3.2
+Flask-HTTPAuth==3.2.1
+Flask-OAuthlib==0.9.3
+Flask-PageDown==0.2.2
+Flask-Babel==0.11.1
+Flask-Cache==0.13.1
+Flask-CacheControl==0.1.2
+Flask-Compress==1.3.2
+Flask-QRcode==1.0
+Markdown>=2.6.6
+Pygments==2.1.3
+bleach==1.4.3
+python-fedora==0.8.0
+python-openid==2.2.5
+python-openid-cla==1.2
+python-openid-teams==1.1
+jieba==0.38
