+++ requirements/dev.txt
@@ -1,0 +1,4 @@
+-r common.txt
+ForgeryPy==0.1
+httpie==0.9.6
+ipython==5.1.0
