+++ requirements/prod.txt
@@ -1,0 +1,4 @@
+-r common.txt
+gunicorn==19.6.0
+uWSGI==2.0.13.1
+python-daemon==2.1.1
